Enhance Task component and add ListTask context
The Task component is improved to import useContext and handle task deletion. The structure of the task and its visual representation are updated. A ListTaskContext is created and implemented in ListTask component to store tasks based on their status (todo, in-progress, done) and provide it to TaskStatusSection component where each section gets its respective tasks. Header component is updated to be more dynamic (resolves #9 #10 #11).