Add drag and drop, confirm before delete, and success confetti
In the Task component, a drag and drop functionality has been added by useSortable from "@dnd-kit/sortable", and a confirm dialog has been added before deleting a task. A confetti effect is triggered whenever a task status is changed to "done". In addition to this, a few code stylings for consistencies and several packages have been installed and updated (resolves #8).